backend/requirements.txt
fastapi
uvicorn
google-generativeai
ezdxf
shapely
python-dotenv
pydantic
requests